refactor: extract reviews logic into dedicated store module
- Create `store/reviews` with slice, thunks, and selectors.
- Remove review-related logic from `store/offer`.
- Register `reviewsSlice` in the root reducer and NameSpace.
- Update `OfferPage` and `CommentForm` to use the new reviews store.
- Move inline styles to CSS classes for better maintainability.
