+++ src/components/comment-form/index.tsx
@@ -1,9 +1,9 @@
 import { useState, ChangeEvent, FormEvent, useEffect, useCallback, memo } from 'react';
 import { useParams } from 'react-router-dom';
 import { useAppDispatch, useAppSelector } from '../../hooks';
-import { postCommentAction } from '../../store/offer/offer-thunks';
-import { dropSendingStatus } from '../../store/offer/offer-slice';
-import { selectSendingStatus, selectSendingError } from '../../store/offer/offer-selectors';
+import { postCommentAction } from '../../store/reviews/reviews-thunks';
+import { dropSendingStatus } from '../../store/reviews/reviews-slice';
+import { selectSendingStatus, selectSendingError } from '../../store/reviews/reviews-selectors';
 import { RequestStatus } from '../../const';
 import './comment-form.css';
 
@@ -135,7 +135,7 @@ function CommentFormComponent(): JSX.Element {
       />
 
       {isError && sendingError && (
-        <div className="reviews__error" style={{ color: 'red', marginBottom: '10px' }}>
+        <div className="reviews__error">
           {sendingError}
         </div>
       )}
