+++ src/store/root-reducer.ts
@@ -4,10 +4,12 @@ import { userSlice } from './user/user-slice';
 import { offersSlice } from './offers/offers-slice';
 import { offerSlice } from './offer/offer-slice';
 import { favoritesSlice } from './favorites/favorites-slice';
+import { reviewsSlice } from './reviews/reviews-slice';
 
 export const rootReducer = combineReducers({
   [NameSpace.User]: userSlice.reducer,
   [NameSpace.Offers]: offersSlice.reducer,
   [NameSpace.Offer]: offerSlice.reducer,
   [NameSpace.Favorites]: favoritesSlice.reducer,
+  [NameSpace.Reviews]: reviewsSlice.reducer,
 });
